+++ src/pipecat/pipeline/task.py
@@ -32,7 +32,6 @@ from pipecat.metrics.metrics import ProcessingMetricsData, TTFBMetricsData
 from pipecat.observers.base_observer import BaseObserver
 from pipecat.pipeline.base_pipeline import BasePipeline
 from pipecat.pipeline.base_task import BaseTask
-from pipecat.pipeline.task_observer import TaskObserver
 from pipecat.processors.frame_processor import FrameDirection, FrameProcessor
 from pipecat.utils.asyncio import BaseTaskManager, TaskManager
 
@@ -184,6 +183,7 @@ class PipelineTask(BaseTask):
         self._idle_timeout_secs = idle_timeout_secs
         self._idle_timeout_frames = idle_timeout_frames
         self._cancel_on_idle_timeout = cancel_on_idle_timeout
+        self._observers = observers
         if self._params.observers:
             import warnings
 
@@ -193,7 +193,7 @@ class PipelineTask(BaseTask):
                     "Field 'observers' is deprecated, use the 'observers' parameter instead.",
                     DeprecationWarning,
                 )
-            observers = self._params.observers
+            self._observers = self._params.observers
         self._finished = False
 
         # This queue receives frames coming from the pipeline upstream.
@@ -228,11 +228,6 @@ class PipelineTask(BaseTask):
         # This task maneger will handle all the asyncio tasks created by this
         # PipelineTask and its frame processors.
         self._task_manager = task_manager or TaskManager()
-
-        # The task observer acts as a proxy to the provided observers. This way,
-        # we only need to pass a single observer (using the StartFrame) which
-        # then just acts as a proxy.
-        self._observer = TaskObserver(observers=observers, task_manager=self._task_manager)
 
         # These events can be used to check which frames make it to the source
         # or sink processors. Instead of calling the event handlers for every
@@ -347,8 +342,6 @@ class PipelineTask(BaseTask):
             self._process_push_queue(), f"{self}::_process_push_queue"
         )
 
-        await self._observer.start()
-
         return self._process_push_task
 
     def _maybe_start_heartbeat_tasks(self):
@@ -367,8 +360,6 @@ class PipelineTask(BaseTask):
             )
 
     async def _cancel_tasks(self):
-        await self._observer.stop()
-
         await self._task_manager.cancel_task(self._process_up_task)
         await self._task_manager.cancel_task(self._process_down_task)
 
@@ -425,7 +416,7 @@ class PipelineTask(BaseTask):
             audio_out_sample_rate=self._params.audio_out_sample_rate,
             enable_metrics=self._params.enable_metrics,
             enable_usage_metrics=self._params.enable_usage_metrics,
-            observer=self._observer,
+            observers=self._observers,
             report_only_initial_ttfb=self._params.report_only_initial_ttfb,
         )
         start_frame.metadata = self._params.start_metadata